+++ R/packages.R
@@ -16,9 +16,13 @@
 #' packages when attaching/loading them but they have not been installed.
 #' @param ... Package names (character vectors, and must always be quoted).
 #' @param install Whether to automatically install packages that are not
-#'   available using \code{\link{install.packages}()}. You are recommended to
-#'   set a CRAN mirror in the global option \code{repos} via
-#'   \code{\link{options}()} if you want to automatically install packages.
+#'   available using \code{\link{install.packages}()}. Besides \code{TRUE} and
+#'   \code{FALSE}, the value of this argument can also be a function to install
+#'   packages (\code{install = TRUE} is equivalent to \code{install =
+#'   install.packages}), or a character string \code{"pak"} (equivalent to
+#'   \code{install = pak::pkg_install}, which requires the \pkg{pak} package).
+#'   You are recommended to set a CRAN mirror in the global option \code{repos}
+#'   via \code{\link{options}()} if you want to automatically install packages.
 #' @param message Whether to show the package startup messages (if any startup
 #'   messages are provided in a package).
 #' @return \code{pkg_attach()} returns \code{NULL} invisibly. \code{pkg_load()}
@@ -37,7 +41,7 @@ pkg_attach = function(
     suppressPackageStartupMessages(base::library(...))
   }
   for (i in c(...)) {
-    if (install && !loadable(i)) pkg_install(i)
+    if (!isFALSE(install) && !loadable(i)) pkg_install(i, install)
     library(i, character.only = TRUE)
   }
 }
@@ -50,8 +54,8 @@ pkg_load = function(..., error = TRUE, install = FALSE) {
   if (n == 0) return(invisible(res))
   for (i in seq_len(n)) {
     res[i] = loadable(p <- pkg[i])
-    if (install && !res[i]) {
-      pkg_install(p); res[i] = loadable(p)
+    if (!isFALSE(install) && !res[i]) {
+      pkg_install(p, install); res[i] = loadable(p)
     }
   }
   if (error && any(!res)) stop('Package(s) not loadable: ', paste(pkg[!res], collapse = ' '))